requirements.txt
aniso8601==8.0.0
ansible==2.8.4
ansible-vault==1.2.0
asn1crypto==0.24.0
attrs==19.1.0
cffi==1.12.3
Click==7.0
cryptography==2.7
docopt==0.6.2
entrypoints==0.3
flake8==3.7.8
Flask==1.1.1
flask-restplus==0.13.0
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.2
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==19.2
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pyparsing==2.4.2
pyrsistent==0.15.4
pytz==2019.2
PyYAML==5.1.2
six==1.12.0
SQLAlchemy==1.3.8
Werkzeug==0.15.6
